requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
requests
pytest
httpx
python-dotenv
sentencepiece
fugashi 
unidic-lite